+++ src/modules/commands/init.ts
@@ -3,16 +3,17 @@ import { blue, green, red, yellow } from "../Colors";
 import { Command, Arg } from "../../types/interfaces";
 import TEV from "../../Main";
 import loadArguments from "../Arguments";
+import fs from "node:fs"
 
 const init = async (tev: TEV, args: string[]): Promise<void> => {
     const command: Command = commands.find((command: Command) => command.name == "version")!;
-    
+
     // ? Arguments executed first, otherwise it loads a default configuration.
-    if(args[0]) {
-        loadArguments(command, args)
+    if (args[0]) {
+        loadArguments(command, args);
     } else {
-        tev.loadConfig();
+        tev.loadInitialConfig();
     }
-}
+};
 
 export default init;