+++ src/modules/commands/run.ts
@@ -6,13 +6,13 @@ import loadArguments from "../Arguments";
 
 const run = async (tev: TEV, args: string[]): Promise<void> => {
     const command: Command = commands.find((command: Command) => command.name == "version")!;
-    
+
     // ? Arguments executed first, otherwise it executes the configurated main file.
-    if(args[0]) {
-        loadArguments(command, args)
+    if (args[0]) {
+        loadArguments(command, args);
     } else {
         await tev.run();
     }
-}
+};
 
 export default run;