+++ src/structures/methods/index.ts
@@ -1,0 +1,3 @@
+// ? Exports every methods.
+export * from "./log";
+export * from "./variables"